fix(CRD): check the logicCode

+++ controllers/alert_reconcile.go
@@ -54,7 +54,7 @@ func createAlertDeployment(cluster *dsv1alpha1.DSAlert) *v1.Deployment {
 			Namespace: cluster.Namespace,
 		},
 		Spec: v1.DeploymentSpec{
-			Replicas: int32Ptr(int32(cluster.Spec.Replicas)),
+			Replicas: int32Ptr(cluster.Spec.Replicas),
 			Selector: &metav1.LabelSelector{
 				MatchLabels: map[string]string{
 					dsv1alpha1.DsAppName: dsv1alpha1.DsAlert,
@@ -71,6 +71,7 @@ func createAlertDeployment(cluster *dsv1alpha1.DSAlert) *v1.Deployment {
 						Name:            dsv1alpha1.DsAlert,
 						Image:           ImageName(cluster.Spec.Repository, cluster.Spec.Version),
 						ImagePullPolicy: corev1.PullIfNotPresent,
+
 						Env: []corev1.EnvVar{
 							{
 								Name:  dsv1alpha1.DataSourceDriveName,
